pyproject.toml
[tool.poetry]
name = "adebot"
version = "0.1.0"
description = ""
authors = ["Jeferson Machado <jeferson_m94@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {extras = ["job-queue"], version = "^20.1"}
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
SQLAlchemy = "^2.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
